// src/app/shared/types.ts
export class NBParameters {
  question: string;
}

export class NBResult {
  accuracy: number;
}

export class Questions {
  district: string;
  facility: string;
  tc: string;
  opdclin: string;
  owner: string;
  sex: string;
  finstat: string;
  weight: number;
}

export class ProbabilityPrediction {
  name: string;
  value: number;
}